sharness: don't remove broker.log with --debug
When using test_under_flux a broker.log file is created to capture
all log messages into a log file. However, there is no way to preserve
this file for successful tests, as other test products can be
preserved with --debug.

Set the broker.log for tests to only be added to cleanup when
--debug has not been used.

Fixes #1288

+++ t/sharness.d/flux-sharness.sh
@@ -65,7 +65,7 @@ test_under_flux() {
     personality=${2:-full}
     log_file="$TEST_NAME.broker.log"
     if test -n "$TEST_UNDER_FLUX_ACTIVE" ; then
-        cleanup rm "${SHARNESS_TEST_DIRECTORY:-..}/$log_file"
+        test "$debug" = "t" || cleanup rm "${SHARNESS_TEST_DIRECTORY:-..}/$log_file"
         flux_module_list > module-list.initial
         cleanup check_module_list
         return
